server: Add a GStreamer 0.10 MJPEG video encoder and use it by default
The GStreamer video encoder is quite primitive and mostly meant to
introduce, test and debug the basic infrastructure.
The main limitation is the lack of rate control: the bitrate is set at
startup and will not react to changes in the network conditions. Still
it should work fine on LANs.