server: Avoid copying the input frame in the GStreamer encoder.
To do so we reference the source bitmap chunks in the GStreamer buffer and rely on the buffer's lifetime being short enough.
Note that we can only avoid copies for the first 1 Mpixels or so. That's because Spice splits larger frames into more chunks and we can fit memory fragments inside a GStreamer buffer. So for those we will avoid copies for the first 3840 KB and copy the rest.