server: Add VP8 support, a video codec preference list and compatibility checks with the Spice client.
The video encoder preferences can be expressed by building a semi-colon
separated list of encoder:codec pairs. For instance
'gstreamer:vp8;spice:mjpeg' to pick first the GStreamer VP8 video
encoder first and used the original MJPEG video encoder one as a
fallback.
The server has a default preference list which can also be selected by
specifying 'auto' as the preferences list.
The client compatibility check relies on the following new capabilities:
* SPICE_DISPLAY_CAP_MULTI_CODEC which denotes a recent client that
supports multiple codecs. This capability is needed to not have to
hardcode that MJPEG is supported. This makes it possible to write
clients that don't support MJPEG.
* SPICE_DISPLAY_CAP_CODEC_XXX, where XXX is a supported codec, for now
MJPEG and VP8.